net: ethernet: asix: Fix asix kernel module autoload
With CONFIG_SPI_AX88796C set as a module, there are issues with
autoloading this kernel module.

For spi devices in OF the modalias exposed to userspace is
spi:<node type>, for the Asix AX88796C driver this is spi:ax88796c

spi:ax88796c

On the other hand, the asix module does not use the same alias:

alias:          of:N*T*Casix,ax88796c*

This happens because the SPI subsystem hardcodes spi:<spi->modalias>, as seen in
drivers/spi/spi.c:

static int spi_uevent(struct device *dev, struct kobj_uevent_env *env)
{
        const struct spi_device *spi = to_spi_device(dev);

        add_uevent_var(env, "MODALIAS=%s%s", SPI_MODULE_PREFIX, spi->modalias);
        return 0;
}

Thus we define the spi id table, add ax88796c to it and have the alias as reported by
the spi driver added so that userspace can correctly autoload the module for the device.

Signed-off-by: Florin Sarbu <florin@resin.io>